Page Title translation capabilities. Switched to the ampersand html entity
I'm saying capabilities because there are too many pages and I can't add
all the translation terms to the domain files at this moment. Some are
going to be picked up because they already exist in the domain files,
though.
I had to enter the ampersand html entity name in order for it to be
displayed correctly.
It's taking into consideration the space before and after that I found
in almost all of the page titles class properties.